+++ packages/cards-editor-config-writer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/cards-editor-config-writer
 
+## 0.5.1
+
+### Patch Changes
+
+-   Updated dependencies [5e0020b]
+    -   @sap-ux/ui5-config@0.28.1
+
+## 0.5.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/ui5-config@0.28.0
+
 ## 0.4.16
 
 ### Patch Changes