+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.2.1
+
+### Patch Changes
+
+-   Updated dependencies [66b88e1]
+    -   @sap-ux/inquirer-common@0.7.1
+    -   @sap-ux/cf-deploy-config-writer@0.3.1
+    -   @sap-ux/project-access@1.30.1
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.1
+    -   @sap-ux/deploy-config-generator-shared@0.1.1
+    -   @sap-ux/fiori-generator-shared@0.12.1
+
+## 0.2.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/deploy-config-generator-shared@0.1.0
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.0
+    -   @sap-ux/cf-deploy-config-writer@0.3.0
+    -   @sap-ux/fiori-generator-shared@0.12.0
+    -   @sap-ux/inquirer-common@0.7.0
+    -   @sap-ux/feature-toggle@0.3.0
+    -   @sap-ux/project-access@1.30.0
+    -   @sap-ux/btp-utils@1.1.0
+
 ## 0.1.60
 
 ### Patch Changes