+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @sap-ux/odata-service-writer
 
+## 0.27.1
+
+### Patch Changes
+
+-   5e0020b: Support backend changes.
+-   Updated dependencies [5e0020b]
+    -   @sap-ux/ui5-config@0.28.1
+    -   @sap-ux/mockserver-config-writer@0.9.1
+    -   @sap-ux/project-access@1.30.1
+
+## 0.27.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/mockserver-config-writer@0.9.0
+    -   @sap-ux/project-access@1.30.0
+    -   @sap-ux/ui5-config@0.28.0
+
 ## 0.26.18
 
 ### Patch Changes