+++ packages/ui5-library-reference-writer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/ui5-library-reference-writer
 
+## 0.2.1
+
+### Patch Changes
+
+-   Updated dependencies [5e0020b]
+    -   @sap-ux/ui5-config@0.28.1
+    -   @sap-ux/project-access@1.30.1
+
+## 0.2.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/project-access@1.30.0
+    -   @sap-ux/ui5-config@0.28.0
+
 ## 0.1.61
 
 ### Patch Changes